Update theme to more closely match Jisc defaults.
The Jisc website seems to use the [62.5% font-size
trick](https://snook.ca/archives/html_and_css/font-size-with-rem).

But adding css to the `html` or `body` elements seems a bit hack-y, and
I don't think we should expect users of this package to have to remember
to do that to get things to look like the Jisc site.

So I've scaled-down all of the font sizes in the typography part of the
theme so they look right on a site that hasn't set font-size: 62.5%.